\docType{methods}
\name{G1,numeric,character-method}
\title{G1,numeric,character-method 
 a method that uses both arguments}
\alias{G1,numeric,character-method}
\usage{\S4method{G1}{numeric,character}(a1,
a2)}
\arguments{
\item{a1}{: of class numeric}
\item{a2}{: of class character}
}
\description{description}
