\description{no Description}
\name{G1}
\alias{G1}
\usage{G1(a1,
a2)}
\title{G1 S4 generic}
\arguments{
\item{a1}{see the method arguments for details}
\item{a2}{see the method arguments for details}
}
\section{Methods}{	\code{\link{G1,numeric,ANY-method}}\cr
	\code{\link{G1,numeric,character-method}}\cr}
